Docker echo update

+++ Automation/deactivate_echo_bot.sh
@@ -1,2 +1,12 @@
 #!/bin/bash
+
+# If no dir is specified, fail
+if [ -z "${ECHOBOT_DIR}" ]; then
+    exit 1
+fi
+
+cd $ECHOBOT_DIR
+
+docker cp echobot:echobot/echobot/log/ persistent_files/
+
 docker rm -f echobot
